Immplement associations with user authentication to accomplish tasks

+++ db/migrate/20150613100540_create_channels.rb
@@ -3,7 +3,7 @@ class CreateChannels < ActiveRecord::Migration[5.0]
     create_table :channels do |t|
       t.string :name, :null => false
       t.string :callsign, :null => false
-      t.decimal :price_per_month, :null => false
+      t.float :price_per_month, :null => false
 
       t.timestamps
     end
